Add files via upload
Assignment specification:
You are required to develop a program in C that will simulate a mathematics quiz
game. The game will include various features, each of which must be incorporated
into your program.
Your program should be menu-driven and must display a main menu when run. The
menu should include the following options:
1. Enter the number of questions to be asked for this round of the quiz
(maximum of 5 questions allowed).
2. Start and play quiz
3. Display the number of questions that were answered (i) correctly and (ii)
incorrectly for this round.
4. Exit Program